Use the proper termination code
This one drove me crazy.   The issue was that
if cmd-Q was used with a dirty window the program
would closed the window after the alert sheet was
processed, but would not terminate.  What I have
to do is cancel termination if I'm going to put up
an alert.  If the result of the alert is to quit
the program I close the window which then triggers
termination.  Or something like that.