Save copy of original file.
If a URL to a save directory is found in user defaults the original file will be linked there in addition to the Trash.
Caveats:  Since I'm using a hard link the save directory must be on the same file system as the original photo.
Link errors will fail silently.   If the directory does not exist the file will not be linked.  If the file already exists in the
save directory the link will fail -- saving multiple times will result in only the 1st original in the save directory.
Multiple copies will still be in the Trash.