More work saving originals to alternate directory
When file can not be trashed exiftool makes a backup by appending _original to the file name.
Attempt to link either the trashed file or the _original file to an optional save folder. If linking
fails fall back to copying to the save folder.

No way to to specify a save folder other than the defaults command at this time.
defaults write org.snafu.GeoTag SaveDirectoryKey /the/path/to/the/save/folder