Optional save folder handling.
The original file will not be linked/copied to the optional save folder if a file of the same name exists in that folder.
If the optional save folder does not exist the link/copy will silently fail.
Also clean up ugly code handling the -overwrite_original option.
Still testing.