Image Backup changes
All Images are now backed up to a save folder which is no longer optional.   The system trash is no longer used.

Image backups are numbered if an image with the same name exists in the backup folder.

A test that the backup exists has been added to work around the filemanager bug were a file copy can fail yet not report an error.

Localizable strings are current for english only.