Use Coord instead of seperate lat/lon
Part way through code clean-up to use Coord (a typealias for CLLocationCoordinate2D) in place of separate latitude and longitude variables.

Not finished.  Not fully tested.

Also, attempt clean up of the aliases left in the sandbox when the program terminates.   Not working on termination, is working on a “clear” request.