Test if exiftool can write the file type
Exiftool can not write all image types.  HEIC files, for example, can be read but not written.  When a file is opened check if exiftool can write the detected file type.  If not add the file as a greyed out entry to the table that can not be updated.  Otherwise GeoTag will allow an update but trigger an error when the file is saved.