Add support for related files
Create an XmpFile class with an extension that conforms to NSFilePresenter.   Use XmpClass with an NSFileCoordinator to read the contents of sidecar files.  I probably have an issue with error handling.  Append a "*" to file names if a sidecar file was found for that file.   The * indicates the sidecar file will be updated instead of the image file.

Question: Why don't I have to use NSFileCoordinator to update the sidecar file?