Unit and User Interface tests
This is a squashed commit of many, many changes to add some test cases to GeoTag.  Many more test cases are needed.  The commit had to be squashed as some of the individual commits added test images too large to upload to github.  A 100+ MB tif image really wasn't needed for testing.

The changes were initially done on a couple of branches which were merged into my local copy of the master branch at different times.