Replace a fatalError with a throwing function
Crash logs from users show that the ImageData init function does sometime fail. Make the initializer throw instead of calling fatalError.   Display an alert for files that could not be processed containing the path of the file and the error.

RUNNING TEST CASES CRASHES XCODE.  THIS MAY BE A CATALINA BETA ISSUE.