Add window to get time zone
Now that I've got this stubbed in I'm not sure if I'll use it.  Still thinking about a good way to handle time zone differences.  I must remember to make a distinction between the time zone of the location where an image was captured and the time zone that was set in the camera.  They may be two different things.