Change Camera TimeZone implemented
Still some cleanup needed.  In short a menu option opens a window to allow a TimeZone to be selected.  The zone will be used with image timestamps to calculate the appropriate time interval needed to compare with Gpx track logs.  If not set the time zone is assumed to be the local time zone of the computer running GeoTag.