Add test image to asset catalog
Add ImageModelImage to make the test image available.  Alas, it doesn't seem to work in View previews which was the reason for adding it in the first place.  Also, the image I reference doesn't contain metadata.  If I ever figure out how to get previews working I'll have to go back and make sure Metadata is being copied to the image.

Remove DispatchQueue processing from open panel code for now.