Map pins
MKPinAnnotationView is deprecated but its replacement, MKMarkerAnnotationView doesn't work as I need.  Or may be I haven't yet figured out how. In any case continue to use the deprecated code for now.

Update the pin location on the map when an image is selected.
Update the image location value when a pin is dragged.
Update both the pin location on the map and the image location values when the map is clicked.