Partial fixes to map clicking and context menu
Clicking on map changes location, but dragging a pin does not cause the location displayed in the table to update until I click away.  View doesn't know anything has changed.

Change tables to use TableRow so I can tie a context menu to each row.  Also add the context menu to the table.  When right clicking on a row the item in that row will be acted upon.  When right clicking on an empty part of the table the selected items will be acted upon.

ContextMenu moved to ImageTable group.