More menu items and action files
Each action is in its own file.  Add actions and menu items for Open, Save, Discard Changes, and Discard Tracks.  Most not yet implemented.

Discovered that the Clear Action is now causing a crash due to a race condition.  Need to figure out where the race is (the related code should be on the main thread) and/or a way around the crash.