Save context for menu items where needed
Some menu actions work on a right clicked image or the selected image, depending upon context.

Add menu items for other functions yet to be added to the code.  Have not yet added all of the associated action functions.   Show In Finder has been implemented.

Work around an issue where is seems the SwiftUI view system holds on to a copy of an ImageModel instance or its ID and tries to access the image that is no longer in the array if images as a result.   Return a dummy "blank" image if the requested image is not in the array.