Map Searching appears to be working
Keep track of map center and altitude to make saving map config easier.  mapCenter is changed and a re center of the view is triggered when a search is performed without any images selected.   When images are selected they are assigned the location of any search result.  Only process the search string upon endEditing, not every change.