Rethink the way I handle selection and context
Do away with a separate "menuContext".  Keyboard shortcuts and main menu will use the current selection.  Right-click and double-click will check if the item is part of a selection.  If so it is made the "most selected" item.  If not the selection will be replaced with the item.  That lets menu actions always work with the current selection -- no special cases to handle.