Make date/time changes a popover, too
Also, dismiss popovers when the Change button is pressed.  No need to show the old timestamp in the popover.  Do not show the popovers if the image is not valid.  Make sure there is a frame for the timestamp so it can be double clicked should a timestamp be empty.