Workaround for crash when scrolling image table.
Scrolling the table would crash when trying to render the DateTime column. The crash was due to the @EnvironmentObject being empty!  To get around this do not rely on the Environment.  Explicitly pass the ViewModel to the column views.

Also get rid of all possible cases where a Published variable in the View Model is bound in a view.  This caused warnings in some cases.  ViewModel published variables are bi-directionally linked to local state variables and the local state is bound in the view.