Menu actions handled in content view
They are not table specific so don't need to be in the Image table view.
Fix pin annotation views so the correct type of pin is always displayed.
Handle proper show/hide of other pins.
Change menu for "Pin view options...".  It is a picker that shows both options using the menu style and lets the user select one.