Create sidecar files for update
When the create sidecar file option is enabled sidecar files will be created for updated images when saving changes.  The original image file will not be touched.  The sidecar file is created with metadata from the original image.  Existing sidecar files will be used when they exist.

Fix bug in saving GPS Timestamps to sidecar files.   The format was date/time/timezone instead of a UTC date/time.