Separate views for table columns.  Again.
For a reason I don't understand table columns are not updated when changed
in some undo invocations when the TableColumn is a Text field dependent upon
a property of an image.  If I make the column a separate vied dependent upon
the image instead the UI is (mostly) updated as expected.