Fix showing other pins.
Massaged the code until the problem went away.  The issue a recursive loop
trying to update map state, but I don't know why checking the status of
a state variable caused the issue in the first place.

Also got rid of otherPins as an observed state.  The are created when needed
from the current selection.