Implement Split H and V views
SwiftUI VSplitView and HSplitView do not remember view positions between
launch. The dividers might also move as the window is resized including
switching between full screen and back.  I partially solved this for the
map and image views, but my implemention had some issues.  The new code
is simple and seems to be working OK.

Position is saved in AppDefaults as a percentage of the window size. Relative
positions are thus maintained as the window is resized.