Fix textfield menu detection
I did not have to do as much as this but making the changes probably
makes for better code AND it let me find the place I was doing the wrong
thing.

I still need to do better enabling, e.g. don't enable if the textfield
is empty.  Need to change the value for that.