Refactor Gpx file processing
GpxTrackLog is a new struct that contains the results of parsing a GPX
file. The struct is Sendable and instances can be used throughout GeoTag
once created.

The Gpx class is now private, used by the above struct init function
to parse a file. An instance of the class does not live longer than
struct init.

16 warnings