Back out the last change to PhotoLibraryController
The change caused a crash (wrong thread) whenever changes would be
saved, making the app useless for updating images in a users Photos
Library.

There is probably a better change, but I want to fix the crash while I'm
looking for the proper way to handle the PhotoLibrary in the app.