First try at handling images with no metadata
Core Graphics can not extract metadata from some raw formats, Fujifilm
compressed raw for example. Show the image name in the table in an
orange color to notify the user that the image has no metadata or it
has metadata that can not be read by GeoTag.  GeoTag/Exiftool can
update the image if the metadata is changed.